最佳迷宫优化以及按键生成

+++ src/include/io/maze_saver.h
@@ -10,7 +10,7 @@ struct SavedMazeInfo {
     QString name;
     QString source;           // "ga", "manual"
     MazeModel maze;
-    int fitness = 0;
+    double fitness = 0.0;
     int dpScore = 0;
     int greedyScore = 0;
     QDateTime timestamp;
@@ -22,7 +22,7 @@ public:
     static bool saveMaze(const QString &path, const MazeModel &maze,
                          const QString &name = QString(),
                          const QString &source = QString(),
-                         int fitness = 0, int dpScore = 0,
+                         double fitness = 0.0, int dpScore = 0,
                          int greedyScore = 0);
 
     static bool loadMaze(const QString &path, SavedMazeInfo &info);
@@ -30,12 +30,12 @@ public:
     static bool saveGAResult(const QString &path,
                              const MazeModel &maze,
                              const OptimizerConfig &config,
-                             int fitness, int dpScore, int greedyScore);
+                             double fitness, int dpScore, int greedyScore);
 
     static QJsonObject mazeToJson(const MazeModel &maze,
                                   const QString &name,
                                   const QString &source,
-                                  int fitness, int dpScore,
+                                  double fitness, int dpScore,
                                   int greedyScore);
 
     static bool loadMazeFromJson(const QJsonObject &json, SavedMazeInfo &info);
